add dockerfiles for Debian buster. Fix build in gcc 7+

+++ plugins/H.263-1998.3/mpi.cxx
@@ -132,8 +132,12 @@ bool MPIList::getNegotiatedMPI( unsigned* width, unsigned* height, unsigned* _fr
   // to the desired one or matches it
   for (i=0; i < MPIs.size(); i++) {
     // we square the value in order to get absolute distances
-    distance = ( abs(MPIs[i].width  - desiredWidth ) *
-                 abs(MPIs[i].height - desiredHeight) );
+    // add static_cast<int>, for gcc 7+ build.
+    // distance = ( fabs(MPIs[i].width  - desiredWidth ) *
+    //              fabs(MPIs[i].height - desiredHeight) );
+
+    distance = ( abs(static_cast<int>(MPIs[i].width  - desiredWidth )) *
+                 abs(static_cast<int>(MPIs[i].height - desiredHeight)) );
 
     if (distance < minDistance) {
       minDistance = distance;
